inputs can control how they're returned
with the :from_given option, a block may be provided that's called with the
given value and any args passed to Inputs#[], and its return value will be the
resulting value (and be cached)

a common use case for this is e.g. "--foo name", where "name" is an identifier
for some Foo object. using this, the Foo object itself can be returned.